WIP: Support socket-based services also on dom0
To be prepared for the rebase on a new revision of PR#3 the changes to
the functions introduced there are intentionally minimal (i.e. no
cleanup).
