+++ agent/qrexec-agent-data.c
@@ -572,226 +572,6 @@ static int process_child_io(libvchan_t *data_vchan,
     return child_process_status;
 }
 
-#define QUBES_MAX_SERVICE_NAME_LEN 31ULL
-#define QUBES_MAX_SERVICE_ARG_LEN 63ULL
-#define QUBES_MAX_SERVICE_DESCRIPTOR_LEN \
-    (QUBES_MAX_SERVICE_NAME_LEN + QUBES_MAX_SERVICE_ARG_LEN + 1ULL)
-#define QUBES_SOCKADDR_UN_MAX_PATH_LEN 1024
-
-#define MAKE_STRUCT(x) \
-    x("/usr/local/etc/qubes-rpc/") \
-    x("/etc/qubes-rpc/")
-static const struct Q {
-    const char *const string;
-    size_t const length;
-} paths[] = {
-#define S(z) { .string = z, .length = sizeof(z) - 1 },
-    MAKE_STRUCT(S)
-#undef S
-};
-static_assert(sizeof("/etc/qubes-rpc/") == 16, "impossible");
-static_assert(sizeof("/usr/local/etc/qubes-rpc/") == 26, "impossible");
-// static_assert(QUBES_MAX_SERVICE_DESCRIPTOR_LEN == 95,
-//         "bad macro definition");
-#define S(z) \
-    static_assert(sizeof(z) + QUBES_MAX_SERVICE_DESCRIPTOR_LEN <= QUBES_SOCKADDR_UN_MAX_PATH_LEN, \
-            "Path too long: " #z);
-MAKE_STRUCT(S)
-#undef S
-#undef MAKE_STRUCT
-
-#if QUBES_MAX_SERVICE_DESCRIPTOR_LEN > PTRDIFF_MAX
-#error impossible
-#endif
-
-static char *parse_qrexec_argument_from_commandline(char *cmdline) {
-    char *end_user;
-    uintptr_t user_len;
-    if (!qrexec_is_fork_server) {
-        end_user = strchr(cmdline, ':');
-        if (!end_user) {
-            fputs("Bad command from dom0: no colon\n", stderr);
-            abort();
-        }
-        end_user++;
-    } else {
-        end_user = cmdline;
-    }
-
-    user_len = (uintptr_t)end_user - (uintptr_t)cmdline;
-    if (user_len > PTRDIFF_MAX) {
-        fputs("absurd user length\n", stderr);
-        abort();
-    }
-
-    if (strncmp(end_user, RPC_REQUEST_COMMAND " ", RPC_REQUEST_COMMAND_LEN + 1) != 0)
-        return NULL;
-    return end_user + RPC_REQUEST_COMMAND_LEN + 1;
-}
-
-static int qubes_connect(int s, const char *buffer, const size_t total_path_length) {
-    // Avoiding an extra copy is NOT worth it!
-#define QUBES_TMP_DIRECTORY "/tmp/qrexec-XXXXXX"
-    char buf[] = QUBES_TMP_DIRECTORY "\0qrexec-socket";
-    struct sockaddr_un remote = { .sun_family = AF_UNIX, .sun_path = { 0 } };
-    static_assert(sizeof buf <= sizeof remote.sun_path,
-                  "maximum path length of AF_UNIX sockets too small");
-    static const size_t path_separator_offset = sizeof QUBES_TMP_DIRECTORY - 1;
-    int result = -1, dummy_errno = -1;
-    socklen_t socket_len;
-    if (total_path_length != strlen(buffer))
-        abort();
-    if (sizeof remote.sun_path <= total_path_length) {
-        // sockaddr_un too small :(
-        if (NULL == mkdtemp(buf))
-            return -1;
-        buf[path_separator_offset] = '/';
-        if (symlink(buffer, buf)) {
-           dummy_errno = errno;
-           goto fail;
-        }
-        memcpy(remote.sun_path, buf, sizeof buf);
-        socket_len = (socklen_t)(offsetof(struct sockaddr_un, sun_path) + sizeof buf);
-    } else {
-        memcpy(remote.sun_path, buffer, total_path_length);
-        remote.sun_path[total_path_length] = '\0';
-        socket_len = (socklen_t)(offsetof(struct sockaddr_un, sun_path) + total_path_length + 1);
-    }
-
-    result = connect(s, (struct sockaddr *) &remote, socket_len);
-    dummy_errno = errno;
-    unlink(buf);
-fail:
-    buf[path_separator_offset] = '\0';
-    rmdir(buf);
-    errno = dummy_errno;
-    return result;
-}
-
-static int execute_qubes_rpc_command(char *cmdline, int *pid, int *stdin_fd, int *stdout_fd, int *stderr_fd) {
-    int s = -1;
-    char *realcmd, *remote_domain;
-    size_t path_length;
-    realcmd = parse_qrexec_argument_from_commandline(cmdline);
-    if (!realcmd) {
-        do_fork_exec(cmdline, pid, stdin_fd, stdout_fd, stderr_fd);
-        return 0;
-    }
-    remote_domain = strchr(realcmd, ' ');
-    if (!remote_domain) {
-        fputs("Bad command from dom0: no remote domain\n", stderr);
-        abort();
-    }
-    *remote_domain++ = '\0';
-    path_length = strlen(realcmd);
-    if (path_length > QUBES_MAX_SERVICE_DESCRIPTOR_LEN) {
-        fputs("Absurdly long command\n", stderr);
-        return -1;
-    }
-    char const *const delimiter = memchr(realcmd, '+', path_length);
-    size_t const service_length = delimiter ?
-        (size_t)(delimiter - realcmd) : path_length;
-    size_t const argument_length = delimiter ?
-        path_length - service_length - 1 : (size_t)-1;
-#define FAIL(msg, err) do { \
-    fputs((msg "\n"), stderr); \
-    errno = (0 ? ((void)(msg), 0) : (err)); \
-    return -1; \
-} while (0)
-
-    if (argument_length + 1 > QUBES_MAX_SERVICE_ARG_LEN + 1) {
-        FAIL("Service argument too long", E2BIG);
-    } else if (!service_length) {
-        FAIL("Service path empty", EINVAL);
-    } else if (service_length > QUBES_MAX_SERVICE_NAME_LEN) {
-        FAIL("Service path too long", ENAMETOOLONG);
-    } else if (path_length > QUBES_MAX_SERVICE_DESCRIPTOR_LEN) {
-        assert(false && "impossible");
-        abort();
-    }
-#undef FAIL
-
-    if ((s = socket(AF_UNIX, SOCK_STREAM, 0)) == -1) {
-        perror("socket");
-        return -1;
-    }
-
-    for (unsigned use_bare_path = 0; use_bare_path < 2; ++use_bare_path) {
-        for (size_t i = 0; i < sizeof(paths)/sizeof(paths[0]); ++i) {
-            char buffer[QUBES_SOCKADDR_UN_MAX_PATH_LEN];
-            size_t const directory_length = paths[i].length;
-            assert(sizeof buffer > path_length);
-            assert(sizeof buffer - path_length > directory_length);
-
-            // The total size of the path (not including NUL terminator).
-            size_t const total_path_length = directory_length +
-                (use_bare_path ? service_length : path_length);
-            memcpy(buffer, paths[i].string, directory_length);
-            memcpy(buffer + directory_length, realcmd, path_length);
-            buffer[total_path_length] = '\0';
-
-            if (!qubes_connect(s, buffer, total_path_length)) {
-                *stdout_fd = *stdin_fd = s;
-                *stderr_fd = -1;
-                *pid = -1;
-                set_nonblock(s);
-                return 0;
-            }
-            switch (errno) {
-            // These cannot happen
-            case EFAULT:       // all of our parameters are in valid memory
-            case EINVAL:       // we passed valid parameters
-            case ENOTSOCK:     // ditto
-            case ENETUNREACH:  // cannot happen for AF_UNIX
-            case EADDRNOTAVAIL:// cannot happen for AF_UNIX
-            case EINPROGRESS:  // this socket is blocking
-            case EBADF:        // `s` was created by a call to `socket()`
-            case EAFNOSUPPORT: // the kernel supports AF_UNIX
-            case EALREADY:     // ditto
-                abort();
-            // These should not happen
-            case EINTR:
-                // Interrupted by a signal - retry
-                --i; // to retry loop iteration
-                continue;
-            case ENOENT:
-            case ENOTDIR:
-                // These errors could also happen with `execve()`
-                break;
-            case EPROTOTYPE:
-            case ETIMEDOUT:
-            case EPERM:
-            case EIO:
-            case EISDIR:
-            case EAGAIN:       // ditto
-#if EAGAIN != EWOULDBLOCK
-            case EWOULDBLOCK:  // ditto
-#endif
-                // Socket server problem and/or misconfiguration.  Fail the whole connection.
-                // (We do not want to fall back because the user may have
-                // overriden behavior for security reasons, and because
-                // fail-fast is much easier to debug).
-                goto fail;
-            case ECONNREFUSED:
-            case EACCES:
-                remote_domain[-1] = ' ';
-                fprintf(stderr, "Executing command as normal: '%s'\n", cmdline);
-                close(s);
-                do_fork_exec(cmdline, pid, stdin_fd, stdout_fd, stderr_fd);
-                return 0;
-            default:
-                /* Unexpected error */
-                break;
-            }
-        }
-    }
-fail:
-    if (s > 0)
-        close(s);
-    return -1;
-}
-
-
 /* Behaviour depends on type parameter:
  *  MSG_SERVICE_CONNECT - create vchan server, pass the data to/from given FDs
  *    (stdin_fd, stdout_fd, stderr_fd), then return remote process exit code
@@ -846,7 +626,7 @@ static int handle_new_process_common(int type, int connect_domain, int connect_p
             send_exit_code(data_vchan, handle_just_exec(cmdline));
             break;
         case MSG_EXEC_CMDLINE:
-            if (execute_qubes_rpc_command(cmdline, &pid, &stdin_fd, &stdout_fd, &stderr_fd) < 0)
+            if (execute_qubes_rpc_command(cmdline, &pid, &stdin_fd, &stdout_fd, &stderr_fd, !qrexec_is_fork_server) < 0)
                 fputs("failed to spawn process\n", stderr);
             fprintf(stderr, "executed %s pid %d\n", cmdline, pid);
             child_process_pid = pid;
